requirements.txt
flask
flask-cors
langchain-community
langchain-openai
langchain
gspread
oauth2client
google-api-python-client
playwright
pytz
hijri-converter
python-dotenv
pydantic
wkhtmltopdf
pdfkit
jinja2